pyproject.toml
[tool.poetry]
name = "ebert"
version = "0.1.0"
description = "AI code review CLI - LLM-agnostic, supports Claude, GPT, Gemini, Ollama"
license = "MIT"
readme = "README.md"
packages = [{include = "ebert", from = "src"}]
keywords = ["ai", "code-review", "llm", "cli", "git", "diff", "claude", "gpt", "gemini", "ollama", "openai", "anthropic"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Quality Assurance",
  "Topic :: Software Development :: Version Control :: Git",
]

[tool.poetry.scripts]
ebert = "ebert.cli:app"

[tool.poetry.dependencies]
python = "^3.11"
typer = ">=0.9.0,<0.12.0"
click = ">=8.1.0,<8.2.0"
rich = "^13.7.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
pyyaml = "^6.0.1"
google-generativeai = {version = "^0.8.0", optional = true}
openai = {version = "^1.0.0", optional = true}
anthropic = {version = "^0.39.0", optional = true}
httpx = "^0.27.0"

[tool.poetry.extras]
gemini = ["google-generativeai"]
openai = ["openai"]
anthropic = ["anthropic"]
all = ["google-generativeai", "openai", "anthropic"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
mypy = "^1.8.0"
ruff = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
indent-width = 2

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]

[tool.mypy]
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
